Updated import routine for Government Spend Pipeline
We have updated the Spend Pipeline routines to deal with the additional
field added by Cabinet Office.  There are some outstanding issues in
the source data that we have queried with Cabinet Office.  For now, we
have coded around them - specifically the field Textbox39 in the import
file.  We think this should be SpendFinancial2019_20.

The import routine for G-Cloud is still not working.  We will get onto
this in the autumn.  In the mean time, you can import the file
G-Cloud-Spend-14-07-14-v2-For-import.csv using phpmyadmin to get the
latest data.

Use the file govspend.sql to create the current schema.  Then run
govspend_cpv.sql to populate the cpv-codes table.